pyproject.toml
[tool.poetry]
name = "aula-de-docker"
version = "0.1.0"
description = ""
authors = ["Gustavo Miranda <gustavocataldomiranda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
